chore: Update ProductListScreen to toggle between Wrap and SingleChildScrollView with Row
This commit modifies the `ProductListScreen` to include a toggle button in the app bar. The toggle button allows the user to switch between using a `Wrap` layout and a `SingleChildScrollView` with a `Row` layout to display the list of products. The `useWrapLayout` boolean variable is used to determine which layout to use. The `ProductList` widget is updated to accept the `useWrapLayout` variable and render the appropriate layout based on its value.

These changes provide flexibility in how the list of products is displayed in the `ProductListScreen` widget.